1) Исправление потенциальных ошибок, связанных с переходом типа данных идентификатора записей справочника на INT8 2) Ф-и для передачи файлов по каналам информационного обмена блоками 3) Базовый тип данных для работы с пространственными данными (GEOMETRY). Пока только записи в таблицах типов атрибутов
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@182 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/general/eioinsertindicator.sql
@@ -1,4 +1,4 @@
-create or replace function eioInsertIndicator(int4, int4, varchar, timestamp, timestamp, timestamp, int4, int4, varchar) returns int4 as
+create or replace function eioInsertIndicator(int8, int4, varchar, timestamp, timestamp, timestamp, int4, int4, varchar) returns int4 as
 $BODY$
 declare
     idRec alias for $1;
